move all cli commands to one place
